Remove data folder
This was getting out of hand really quickly so decided to keep everything only in a database.

When the need comes (for example to document description, syntax, examples, bugs and polyfills) this folder will come back. Maybe in a slightly different format to limit the amount of files and folders.