Better use of Promises in API controller and improved put/patch actions
Turns out most mongoose methods already return a Promise if no callback is given, so use those.

Improved the patch action by using the `mongoose-json-patch` module, which is a mongoose plugin based on the `fast-json-patch` module. I had to patch it to return a Promise if no callback is given.

Known issue with the put action... `property.update(requestBody);` does not overwrite the whole document. Setting `overwrite: true` should work but gave an error.