Change way to store browser detects
Because using an array doesn't make any sense. Can get duplicate browser versions easily.

In MongoDB you can't use a dot in a fieldname, so can't have browser version in just one field name. Decided on something like this:

    "browsers": {
        "chromefox": {
            "12": {
                // Chromefox 12 doesn't support this property :(
                "in": false,
                "own": false,
                "1": {
                    // Chromefox 12.1 supports this property!
                    "in": false,
                    "own": false
                }
            }
        }
    }

That way it's easy to query for versions, e.g. `browsers.chromefox.12.1`. But how to define that with mongoose? :D